move relativePath to util/strings
